st0102: add TransportStreamId reading, improve overall test coverage
This implements the (ST0102.10) Transport Stream Id key for read only. That isn't supposed to be
used any more, so no public constructor or Builder support.

Remove Builder support for ItemDesignatorId (which is also obsolete, we should only support it
for reading).

Fix display name for SCI/SHI Information (to title case, was "information") to be consistent
with the other labels.

Improve test coverage across ST0102.