st1201: implement encoding / decoding of 5-7 byte values, fix overflow case (#162)
The python code is used to generate 2128 test vectors, and the new tests check forward
and reverse conversion. That showed problems with cases where the high bit is set
(e.g. 0x800000, which is a normal encoded value). That can be fixed by using a
larger type, except for the 8-byte cases, where I added a special-case check.