chore: remove TODO comment
I had a look at alternatives to String for `LocalDate.parse()`, and while it can potentially take
any CharSequence, those require char[] and we're working from UTF-8 byte[].

In addition, the implementation of `LocalDate.parse()` ends up calling into a method that calls
`toString()` on the input text, so its not going to save us anything.

So just remove the comment.